backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.8.1
celery==5.4.0
redis==5.2.1
boto3==1.37.38
psycopg[binary]==3.2.6
numpy>=2.2,<3
scipy>=1.15,<2
pillow==11.1.0
python-multipart==0.0.20
httpx==0.28.1
tenacity==9.0.0
openai==1.75.0
instructor==1.7.9
prometheus-client==0.21.1
portalocker==3.1.1
zarr>=2.18,<3
opensimplex==0.4.5.1
pyfastnoiselite==0.0.7
pyflwdir>=0.5.10
tbb>=2021.0
shapely>=2.0,<3
geopandas>=1.0,<2
scikit-image>=0.22,<1
rasterio>=1.3,<2
GeoAlchemy2>=0.14,<1
pytest>=8.0,<9
